Correcting notations: left to right
While it may sound strange, had been using left-to-right notation: 1 is the first step, 2 the second, 4 the third, etc. In this way, hex and oct notations go in the same direction as the actual beats. So, `0x333:12` is a pattern of three “Scotch snaps” in a row and `0o5555:12` is a measure of “12/8 Shuffle”.